Solve la-hadamard-product

+++ study-plans/math-linear-algebra/la-hadamard-product/la-hadamard-product.py
@@ -4,7 +4,6 @@ def hadamard_product(A, B):
     """
     Returns: ndarray, the element-wise product A * B.
     """
-    A = np.asarray(A)
-    B = np.asarray(B)
-    return A * B
-            
+    A = np.asarray(A, dtype=float)
+    B = np.asarray(B, dtype=float)
+    return A*B
